requirements.txt
Django==1.10
django-mptt==0.8.6
djangorestframework==3.4.6
httpie==0.9.6
Pillow==3.3.1
Pygments==2.1.3
requests==2.11.1
